requirements.txt
dotenv
google-genai
openpyxl
pyyaml
requests